# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ZIAY API
  version: 1.0.0
  description: |
    Comercio Conversacional + Atribución Inteligente.

    Multi-tenant conversational-commerce platform with agent-driven checkout
    (AP2 / UCP / ACP / MCP protocol trinity), ad attribution, compliance
    (Ley 1581 / Ley 2573 / Ley 1480 / DIAN), and a wallet/commission layer
    for traffickers + marketplaces.

    All tenant-scoped endpoints require a session cookie (`sessionAuth`)
    unless marked otherwise. UCP / ACP / webhook endpoints use `bearerAuth`
    or are public.

  contact:
    email: api@ziay.co
  license:
    name: Proprietary
servers:
  - url: https://ziay.co
    description: Production
  - url: http://localhost:3000
    description: Development
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth v4 session cookie (set by /api/auth/[...nextauth]/callback/credentials).
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed W3C Verifiable Credential or bearer token (UCP/AP2/ACP).
    cronAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Bearer $CRON_SECRET — shared secret for /api/compliance/retention/cron.'
    webhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: HMAC-SHA256 signature from the gateway (Meta/Stripe/Wompi/etc.).
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
...
  parameters:
    TenantId:
      name: tenantId
      in: query
      required: true
      schema:
        type: string
      description: Tenant slug (e.g. "saramantha").
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor for cursor-based pagination (id of the last item on the previous page).
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
      description: Page size (max 100).
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - sessionAuth: []
paths:
...
  /api/catalog/products:
    get:
      summary: List catalog products
      description: Search + pagination over the tenant's product catalog.
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: q
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated product list.
  /api/catalog/send-to-chat:
    post:
      summary: Send product card to chat
      description: Push a product card into an active conversation (chat-up sell flow).
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Send confirmation.
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/ap2/mandates:
    get:
      summary: List mandates
      description: List AP2 mandates for a tenant, optionally filtered by userId/type/status.
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: userId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [intent, cart, payment]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, consumed, revoked]
      responses:
        '200':
          description: Mandate list.
    post:
      summary: Create Intent Mandate
      description: Create a signed W3C VC Intent Mandate — first link in the Intent → Cart → Payment chain.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tenantId:
                  type: string
                userId:
                  type: string
                maxAmount:
                  type: number
                currency:
                  type: string
                  default: COP
                categoryLimits:
                  type: object
                expiresAt:
                  type: string
                  format: date-time
                purpose:
                  type: string
      responses:
        '201':
          description: Created mandate + signed VC.
...
  /api/buyer-behavior:
    get:
      summary: List buyer-behavior records
      description: BuyerBehavior records + risk-level stats.
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: '{ behaviors, stats }'
    post:
      summary: Upsert buyer-behavior record
      description: Upsert by (tenantId, phone). When riskLevel is 'high_risk'/'blacklist', also creates a BehaviorAlert.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: '{ behavior, alert }'
...